+++ visualization/app/app.module.ts
@@ -39,7 +39,7 @@ import { CodeChartaModule } from "./codeCharta/codeCharta.module"
 import { UpdateVisibleTopLabelsEffect } from "./codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect"
 import { ResetSelectedEdgeMetricWhenItDoesntExistAnymoreEffect } from "./codeCharta/state/effects/resetSelectedEdgeMetricWhenItDoesntExistAnymore/resetSelectedEdgeMetricWhenItDoesntExistAnymore.effect"
 import { LinkColorMetricToHeightMetricEffect } from "./codeCharta/state/effects/linkColorMetricToHeightMetric/linkColorMetricToHeightMetric.effect"
-import { UpdateAttributeTypesEffect } from "./codeCharta/state/effects/updateAttributeTypes/updateAttributeTypes.effect"
+import { UpdateFileSettingsEffect } from "./codeCharta/state/effects/updateFileSettings/updateFileSettings.effect"
 
 @NgModule({
 	imports: [
@@ -58,8 +58,8 @@ import { UpdateAttributeTypesEffect } from "./codeCharta/state/effects/updateAtt
 			AutoFitCodeMapEffect,
 			UpdateVisibleTopLabelsEffect,
 			LinkColorMetricToHeightMetricEffect,
-			UpdateAttributeTypesEffect,
-			ResetSelectedEdgeMetricWhenItDoesntExistAnymoreEffect
+			ResetSelectedEdgeMetricWhenItDoesntExistAnymoreEffect,
+			UpdateFileSettingsEffect
 		]),
 		SliderModule,
 		AttributeSideBarModule,